requirements.txt
python-telegram-bot
gunicorn
openai
python-dotenv
python-telegram-bot[webhooks]
